refactor: implement arch issues 693-696
- #695: move format_initial_prompt + format_vocabulary_prompt_capped
  from pipeline.rs to dictionary/vocabulary/mod.rs; export via
  crate::dictionary::vocabulary directly

- #696: emit dictation:meeting-append-failed from stop_dictation
  spawn on repo error; add frontend event constant, appendFailedNotice
  state in meeting-sessions.svelte.ts, listener in AppLifecycle.svelte,
  and warning banner in +page.svelte

- #693: extract MeetingAutostartSection.svelte (autostart IPC + markup)
  and DiarizerModelSection.svelte (diarizer toggle + model lifecycle)
  from MeetingTab.svelte; MeetingTab becomes thin coordinator owning
  only app-classifier override state

- #694: create src/lib/state/models.svelte.ts with ModelFetch type,
  modelFetch reactive state, loadModels/downloadModel/cancelDownload/
  removeModel; SettingsPanel imports models module, selectModel + event
  listeners stay in SettingsPanel
